+++ management/samples/hello-world-entity/pom.xml
@@ -39,8 +39,13 @@
   <dependencies>
     <!-- inside jar -->
     <dependency>
-      <groupId>org.terracotta</groupId>
-      <artifactId>voltron-proxy</artifactId>
+      <groupId>org.terracotta.voltron.proxy</groupId>
+      <artifactId>voltron-proxy-client</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.terracotta.voltron.proxy</groupId>
+      <artifactId>voltron-proxy-server</artifactId>
       <version>${project.version}</version>
     </dependency>
 
@@ -120,7 +125,9 @@
               <!-- we have to make sure no voltron api comes in the shaded jar -->
               <artifactSet>
                 <includes>
-                  <include>org.terracotta:voltron-proxy</include>
+                  <include>org.terracotta.voltron.proxy:voltron-proxy-common</include>
+                  <include>org.terracotta.voltron.proxy:voltron-proxy-client</include>
+                  <include>org.terracotta.voltron.proxy:voltron-proxy-server</include>
                 </includes>
               </artifactSet>
             </configuration>
