+++ management/tms-entity/tms-entity-client/pom.xml
@@ -31,6 +31,11 @@
 
   <dependencies>
     <dependency>
+      <groupId>org.terracotta.voltron.proxy</groupId>
+      <artifactId>voltron-proxy-client</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>org.terracotta.management</groupId>
       <artifactId>tms-entity-common</artifactId>
       <version>${project.version}</version>
