+++ pom.xml
@@ -34,7 +34,7 @@
 
   <modules>
     <module>offheap-resource</module>
-    <module>proxy</module>
+    <module>voltron-proxy</module>
     <module>concurrent-map-entity/common</module>
     <module>concurrent-map-entity/client</module>
     <module>concurrent-map-entity/server</module>
