+++ voltron-proxy/voltron-proxy-client/src/main/java/org/terracotta/voltron/proxy/client/ServerMessageAware.java
@@ -13,16 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terracotta.management.entity.management.client;
+package org.terracotta.voltron.proxy.client;
 
-import org.terracotta.connection.entity.Entity;
-import org.terracotta.management.entity.management.ManagementAgent;
-import org.terracotta.management.model.message.ManagementCallMessage;
-import org.terracotta.voltron.proxy.client.ServerMessageAware;
+import org.terracotta.voltron.proxy.MessageListener;
 
 /**
- * @author Mathieu Carbou
+ * @author Alex Snaps
  */
-public interface ManagementAgentEntity extends ManagementAgent, Entity, ServerMessageAware<ManagementCallMessage> {
+public interface ServerMessageAware<T> {
 
+  void registerListener(MessageListener<T> listener);
 }
