+++ voltron-proxy/voltron-proxy-common/src/main/java/org/terracotta/voltron/proxy/Codec.java
