+++ voltron-proxy/voltron-proxy-common/src/test/java/org/terracotta/voltron/proxy/CommonProxyFactoryTest.java
